move codes only for cat command to cat.go and cat_test.go
